requirements.txt
django==4.2
django-admin
django-environ
django-cors-headers
django-webpack-loader
djangorestframework

psycopg
